Refactors App component for streamlined UI
Simplifies the App component by removing form state management.

Moves state management and form handling logic into the WearableList component, enhancing modularity.
Removes the header prop for adding a wearable.
Adds a hero section to the home screen.
