+++ src/App.tsx
@@ -1,54 +1,26 @@
-import { useState } from "react";
 import WearableList from "./components/WearableList";
 import Header from "./components/Header";
-import AddWearableForm from "./components/AddWearableForm";
-import useWearablesStore from "./hooks/useWearablesStore";
-import type { Wearable } from "./lib/types";
 
 function App() {
-  const [wearables, { addWearable, updateWearable }] = useWearablesStore();
-  const [showAddForm, setShowAddForm] = useState(false);
-  const [editingWearable, setEditingWearable] = useState<Wearable | null>(null);
-
-  const handleAddNewWearable = () => {
-    setEditingWearable(null);
-    setShowAddForm(true);
-  };
-
-  const handleEditWearable = (wearable: Wearable) => {
-    setEditingWearable(wearable);
-    setShowAddForm(true);
-  };
-
-  const handleFormSubmit = (wearableData: Omit<Wearable, "id">) => {
-    if (editingWearable) {
-      updateWearable(editingWearable.id, wearableData);
-    } else {
-      addWearable(wearableData);
-    }
-    handleCloseForm();
-  };
-
-  const handleCloseForm = () => {
-    setShowAddForm(false);
-    setEditingWearable(null);
-  };
-
   return (
     <div className="min-h-screen bg-background">
-      <Header onAddWearable={handleAddNewWearable} />
-      <main className="px-4 py-8">
-        <WearableList onEditWearable={handleEditWearable} />
-      </main>
+      <Header />
 
-      {/* Add/Edit Wearable Form */}
-      {showAddForm && (
-        <AddWearableForm
-          onSubmit={handleFormSubmit}
-          onCancel={handleCloseForm}
-          editingWearable={editingWearable}
-        />
-      )}
+      {/* Hero Section */}
+      <section className="bg-background py-16 px-4 sm:px-6 lg:px-8">
+        <div className="max-w-4xl mx-auto text-center">
+          <h1 className="text-4xl sm:text-5xl lg:text-6xl font-bold text-foreground mb-6">
+            AI Wearables Collection
+          </h1>
+          <p className="text-lg sm:text-xl text-muted-foreground max-w-2xl mx-auto">
+            Discover and manage the latest AI-powered wearable devices
+          </p>
+        </div>
+      </section>
+
+      <main className="w-full px-4 py-8">
+        <WearableList />
+      </main>
     </div>
   );
 }
